Update linked list destroy and create methods
Same story that with the node, I was passing a copy of the pointer to
the linked list instead of the actual pointer, so when destroying I
could reach the address in the pointer, but then I left a dangling
pointer due to pointing the copy to NULL, but not the actual pointer.

Similar with create, as node's destroy changed now I need to pass a
pointer to the node, not a copy of the pointer.